Configurando arquivo teacher.js
- Configurando o arquivo teacher.je em models para realizar a inserção
e a busca dos dados dos professores.
