+++ src/app/models/teacher.js
@@ -1,0 +1,40 @@
+const { age, date } = require('../../lib/utils')
+const db = require('../../config/db')
+
+module.exports = {
+    all(callback) {
+        db.query(`SELECT * FROM teachers`, function(err, results) {
+            if(err) return res.send('Erro no banco de dados!')
+
+            callback(results.rows)
+        })
+    },
+    create(data, callback) {
+        const query = `
+            INSERT INTO teachers (
+                name,
+                avatar_url,
+                sexo,
+                disciplinas,
+                birth,
+                created_at
+            ) VALUES ($1, $2, $3, $4, $5, $6)
+            RETURNING id
+        `
+
+        const values = [
+            data.name,
+            data.avatar_url,
+            data.sexo,
+            data.disciplinas,
+            date(data.birth).iso,
+            date(Date.now()).iso
+        ]
+
+        db.query(query, values, function(err, results) {
+            if(err) return res.send('Erro no banco de dados!')
+            
+            callback(results.rows[0])
+        })
+    }
+}
